+++ seed/python-sdk/bearer-token-environment-variable/src/seed/client.py
@@ -11,6 +11,26 @@ from .resources.service.client import AsyncServiceClient, ServiceClient
 
 
 class SeedBearerTokenEnvironmentVariable:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+
+    Parameters:
+        - base_url: str. The base url to use for requests from the client.
+
+        - api_key: typing.Optional[typing.Union[str, typing.Callable[[], str]]].
+
+        - timeout: typing.Optional[float]. The timeout to be used, in seconds, for requests by default the timeout is 60 seconds.
+
+        - httpx_client: typing.Optional[httpx.Client]. The httpx client to use for making requests, a preconfigured client is used by default, however this is useful should you want to pass in any custom httpx configuration.
+    ---
+    from seed.client import SeedBearerTokenEnvironmentVariable
+
+    client = SeedBearerTokenEnvironmentVariable(
+        api_key="YOUR_API_KEY",
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,
@@ -32,6 +52,26 @@ class SeedBearerTokenEnvironmentVariable:
 
 
 class AsyncSeedBearerTokenEnvironmentVariable:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+
+    Parameters:
+        - base_url: str. The base url to use for requests from the client.
+
+        - api_key: typing.Optional[typing.Union[str, typing.Callable[[], str]]].
+
+        - timeout: typing.Optional[float]. The timeout to be used, in seconds, for requests by default the timeout is 60 seconds.
+
+        - httpx_client: typing.Optional[httpx.AsyncClient]. The httpx client to use for making requests, a preconfigured client is used by default, however this is useful should you want to pass in any custom httpx configuration.
+    ---
+    from seed.client import AsyncSeedBearerTokenEnvironmentVariable
+
+    client = AsyncSeedBearerTokenEnvironmentVariable(
+        api_key="YOUR_API_KEY",
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,