+++ seed/python-sdk/custom-auth/src/seed/client.py
@@ -9,6 +9,26 @@ from .resources.custom_auth.client import AsyncCustomAuthClient, CustomAuthClien
 
 
 class SeedCustomAuth:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+
+    Parameters:
+        - base_url: str. The base url to use for requests from the client.
+
+        - custom_auth_scheme: str.
+
+        - timeout: typing.Optional[float]. The timeout to be used, in seconds, for requests by default the timeout is 60 seconds.
+
+        - httpx_client: typing.Optional[httpx.Client]. The httpx client to use for making requests, a preconfigured client is used by default, however this is useful should you want to pass in any custom httpx configuration.
+    ---
+    from seed.client import SeedCustomAuth
+
+    client = SeedCustomAuth(
+        custom_auth_scheme="YOUR_CUSTOM_AUTH_SCHEME",
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,
@@ -26,6 +46,26 @@ class SeedCustomAuth:
 
 
 class AsyncSeedCustomAuth:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+
+    Parameters:
+        - base_url: str. The base url to use for requests from the client.
+
+        - custom_auth_scheme: str.
+
+        - timeout: typing.Optional[float]. The timeout to be used, in seconds, for requests by default the timeout is 60 seconds.
+
+        - httpx_client: typing.Optional[httpx.AsyncClient]. The httpx client to use for making requests, a preconfigured client is used by default, however this is useful should you want to pass in any custom httpx configuration.
+    ---
+    from seed.client import AsyncSeedCustomAuth
+
+    client = AsyncSeedCustomAuth(
+        custom_auth_scheme="YOUR_CUSTOM_AUTH_SCHEME",
+        base_url="https://yourhost.com/path/to/api",
+    )
+    """
+
     def __init__(
         self,
         *,