+++ seed/python-sdk/trace/src/seed/client.py
@@ -17,6 +17,32 @@ from .resources.v_2.client import AsyncV2Client, V2Client
 
 
 class SeedTrace:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+
+    Parameters:
+        - base_url: typing.Optional[str]. The base url to use for requests from the client.
+
+        - environment: SeedTraceEnvironment. The environment to use for requests from the client. from .environment import SeedTraceEnvironment
+
+                                             Defaults to SeedTraceEnvironment.PROD
+
+        - x_random_header: typing.Optional[str].
+
+        - token: typing.Optional[typing.Union[str, typing.Callable[[], str]]].
+
+        - timeout: typing.Optional[float]. The timeout to be used, in seconds, for requests by default the timeout is 60 seconds.
+
+        - httpx_client: typing.Optional[httpx.Client]. The httpx client to use for making requests, a preconfigured client is used by default, however this is useful should you want to pass in any custom httpx configuration.
+    ---
+    from seed.client import SeedTrace
+
+    client = SeedTrace(
+        x_random_header="YOUR_X_RANDOM_HEADER",
+        token="YOUR_TOKEN",
+    )
+    """
+
     def __init__(
         self,
         *,
@@ -44,6 +70,32 @@ class SeedTrace:
 
 
 class AsyncSeedTrace:
+    """
+    Use this class to access the different functions within the SDK. You can instantiate any number of clients with different configuration that will propogate to these functions.
+
+    Parameters:
+        - base_url: typing.Optional[str]. The base url to use for requests from the client.
+
+        - environment: SeedTraceEnvironment. The environment to use for requests from the client. from .environment import SeedTraceEnvironment
+
+                                             Defaults to SeedTraceEnvironment.PROD
+
+        - x_random_header: typing.Optional[str].
+
+        - token: typing.Optional[typing.Union[str, typing.Callable[[], str]]].
+
+        - timeout: typing.Optional[float]. The timeout to be used, in seconds, for requests by default the timeout is 60 seconds.
+
+        - httpx_client: typing.Optional[httpx.AsyncClient]. The httpx client to use for making requests, a preconfigured client is used by default, however this is useful should you want to pass in any custom httpx configuration.
+    ---
+    from seed.client import AsyncSeedTrace
+
+    client = AsyncSeedTrace(
+        x_random_header="YOUR_X_RANDOM_HEADER",
+        token="YOUR_TOKEN",
+    )
+    """
+
     def __init__(
         self,
         *,